+++ test/gtest/common/test.h
@@ -174,8 +174,8 @@ protected:
     unsigned                        m_num_log_handlers_before;
 
     static std::set<int>            m_prev_open_fds;
-    static int                      m_consecutive_fd_increases;
-    static int                      m_total_fd_increases;
+    static size_t                   m_consecutive_fd_increases;
+    static size_t                   m_total_fd_increases;
     static pthread_mutex_t          m_logger_mutex;
     static unsigned                 m_total_errors;
     static unsigned                 m_total_warnings;
